# src/test/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: api-checklist
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /checklists/{checklistId}/phases:
    get:
      tags:
      - Phase resources
      summary: Fetch all phases in a checklist
      operationId: fetchChecklistPhases
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phase"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phase"
    post:
      tags:
      - Phase resources
      summary: Create phase in a checklist
      operationId: createChecklistPhase
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhaseCreateRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "201":
          description: Successful Operation
          headers:
            Location:
              style: simple
              schema:
                type: string
...
  /employee-checklists/{employeeChecklistId}/phases/{phaseId}:
    patch:
      tags:
      - Employee checklist resources
      summary: Update of all tasks in a phase
      description: Bulk update of sent in attributes for all tasks in phase
      operationId: updateAllTasksInPhase
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeChecklistPhaseUpdateRequest"
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistPhase"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistPhase"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /checklists/{checklistId}/phases/{phaseId}:
    get:
      tags:
      - Phase resources
      summary: Fetch phase in a checklist
      operationId: fetchChecklistPhase
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Phase"
    delete:
      tags:
      - Phase resources
      summary: Delete phase in a checklist
      operationId: deleteChecklistPhase
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: Successful Operation
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Phase resources
      summary: Update phase in a checklist
      operationId: updateChecklistPhase
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhaseUpdateRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Phase"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
...
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    QuestionType:
      type: string
      description: The question type of the task
      writeOnly: true
      enum:
      - YES_OR_NO
      - YES_OR_NO_WITH_TEXT
      - COMPLETED_OR_NOT_RELEVANT
      - COMPLETED_OR_NOT_RELEVANT_WITH_TEXT
...
    RoleType:
      type: string
      description: The role type of the task
      readOnly: true
      enum:
      - EMPLOYEE
      - MANAGER
...
    Permission:
      type: string
      description: The permission needed to administrate the phase
      writeOnly: true
      enum:
      - SUPERADMIN
      - ADMIN
    PhaseCreateRequest:
      required:
      - name
      - permission
      - roleType
      - sortOrder
      type: object
      properties:
        name:
          type: string
          description: The name of the phase
          writeOnly: true
          example: Första veckan
        bodyText:
          type: string
          description: The body text of the phase
          writeOnly: true
          example: Detta är en beskrivning av vad som ska göras under första veckan
        timeToComplete:
          type: string
          description: The time to complete the phase
          writeOnly: true
          example: P1M
        roleType:
          $ref: "#/components/schemas/RoleType"
        permission:
          $ref: "#/components/schemas/Permission"
        sortOrder:
          type: integer
          description: The sort order of the phase
          format: int32
          writeOnly: true
          example: 1
      description: Model for phase create request
...
    Phase:
      type: object
      properties:
        id:
          type: string
          description: The id of the phase
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        name:
          type: string
          description: The name of the phase
          readOnly: true
          example: Första veckan
        bodyText:
          type: string
          description: The body text of the phase
          readOnly: true
          example: Detta är en beskrivning av vad som ska göras under första veckan
        timeToComplete:
          type: string
          description: The time to complete the phase
          readOnly: true
          example: P1M
        roleType:
          $ref: "#/components/schemas/RoleType"
        permission:
          $ref: "#/components/schemas/Permission"
        sortOrder:
          type: integer
          description: The sort order of the phase
          format: int32
          readOnly: true
          example: 1
        tasks:
          type: array
          description: Tasks in the phase
          readOnly: true
          items:
            $ref: "#/components/schemas/Task"
        created:
          type: string
          description: The created date and time of the phase
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
        updated:
          type: string
          description: The last update date and time of the phase
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
      description: Model for a phase
    Task:
      type: object
      properties:
        id:
          type: string
          description: The id of the task
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        heading:
          type: string
          description: The heading of the task
          readOnly: true
          example: Bjud på fika
        text:
          type: string
          description: The body text of the task
          readOnly: true
          example: Detta är en beskrivning av ett uppdrag
        sortOrder:
          type: integer
          description: The sort order of the task
          format: int32
          readOnly: true
          example: 1
        roleType:
          $ref: "#/components/schemas/RoleType"
        questionType:
          $ref: "#/components/schemas/QuestionType"
        permission:
          $ref: "#/components/schemas/Permission"
        created:
          type: string
          description: The date and time the task was created
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
        updated:
          type: string
          description: The date and time the task was last updated
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
      description: Task model
...
    FulfilmentStatus:
      type: string
      description: The status of the task fulfilment
      nullable: true
      writeOnly: true
      enum:
      - EMPTY
      - "TRUE"
      - "FALSE"
    EmployeeChecklistTask:
      type: object
      properties:
        id:
          type: string
          description: The id of the task
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        heading:
          type: string
          description: The heading of the task
          readOnly: true
          example: Bjud på fika
        text:
          type: string
          description: The body text of the task
          readOnly: true
          example: Detta är en beskrivning av ett uppdrag
        sortOrder:
          type: integer
          description: The sort order for the task
          format: int32
          readOnly: true
        roleType:
          $ref: "#/components/schemas/RoleType"
        questionType:
          $ref: "#/components/schemas/QuestionType"
        customTask:
          type: boolean
          description: Tells if the task is only applies to the current checklist
            or not
          readOnly: true
        responseText:
          type: string
          description: The task response text
          readOnly: true
          example: Jag har bjudit på fika
        fulfilmentStatus:
          $ref: "#/components/schemas/FulfilmentStatus"
        updated:
          type: string
          description: The date and time the task was last updated
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
      description: Model for a employee checklist task
    EmployeeChecklistPhaseUpdateRequest:
      type: object
      properties:
        tasksFulfilmentStatus:
          $ref: "#/components/schemas/FulfilmentStatus"
      description: Model for update request on an employee checklist phase
    EmployeeChecklistPhase:
      type: object
      properties:
        id:
          type: string
          description: The id of the phase
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        name:
          type: string
          description: The name of the phase
          readOnly: true
          example: Första veckan
        bodyText:
          type: string
          description: The body text of the phase
          readOnly: true
          example: Detta är en beskrivning av vad som ska göras under första veckan
        timeToComplete:
          type: string
          description: The time to complete the phase
          readOnly: true
          example: P1M
        roleType:
          $ref: "#/components/schemas/RoleType"
        sortOrder:
          type: integer
          description: The sort order for the phase
          format: int32
          readOnly: true
          example: 1
        tasks:
          type: array
          description: Tasks in the phase
          readOnly: true
          items:
            $ref: "#/components/schemas/EmployeeChecklistTask"
      description: Model for a employee checklist phase
...
    PhaseUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: The name of the phase
          writeOnly: true
          example: Första veckan
        bodyText:
          type: string
          description: The body text of the phase
          writeOnly: true
          example: Detta är en beskrivning av vad som ska göras under första veckan
        timeToComplete:
          type: string
          description: The time to complete the phase
          writeOnly: true
          example: P1M
        roleType:
          $ref: "#/components/schemas/RoleType"
        permission:
          $ref: "#/components/schemas/Permission"
        sortOrder:
          type: integer
          description: The sort order of the phase
          format: int32
          writeOnly: true
          example: 1
      description: Model for phase update request
...
  securitySchemes: {}